# manifest: small optimizations to NewL0Sublevels