- instead of maintaining the slices in levelFiles in
  sorted order, which is O(n^2), the code appends and
  sorts at the end.
- the search for maxIntervalIndex is pruned to start
  at minIntervalIndex.

Benchmark results (the K benchmark uses the recent
customer issue MANIFEST that I can't include here):

```
name                    old time/op    new time/op    delta
L0SublevelsInit-16        6.92ms ± 0%    6.69ms ± 0%   ~     (p=1.000 n=1+1)
L0SublevelsInitK-16     4.57s ± 0%     4.31s ± 0%   ~     (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
L0SublevelsInit-16        3.03MB ± 0%    3.03MB ± 0%   ~     (p=1.000 n=1+1)
L0SublevelsInitK-16    1.20GB ± 0%    1.20GB ± 0%   ~     (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
L0SublevelsInit-16         33.8k ± 0%     33.8k ± 0%   ~     (p=1.000 n=1+1)
L0SublevelsInitK-16     4.33M ± 0%     4.33M ± 0%   ~     (p=1.000 n=1+1)
```